# openapi.yaml
openapi: 3.0.3
info:
  title: Hateblog Backend API
  description: |
    hateblog リニューアル版のバックエンド API 仕様書。
    はてなブックマークのエントリー情報を提供するREST APIです。
  version: 1.0.0
  contact:
    name: Hateblog Team
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://hateblog.jp/api/v1
    description: 本番環境
security:
  - ApiKeyAuth: []
...
paths:
  /entries/new:
    get:
      tags:
        - entries
      summary: 新着順エントリー一覧取得
      description: |
        指定日付のエントリーを新着順（posted_at DESC）で取得します。
        ブックマーク件数での閾値フィルタリングが可能です。
      operationId: getNewEntries
      parameters:
        - name: date
          in: query
          description: 取得対象日付（YYYYMMDD形式）
          required: true
          schema:
            type: string
            pattern: '^\d{8}$'
            example: "20250105"
        - name: min_users
          in: query
          description: 最低ブックマーク件数（5/10/50/100/500/1000など）
          required: false
          schema:
            type: integer
            minimum: 0
            default: 5
            example: 10
        - name: limit
          in: query
          description: 取得件数
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
            example: 25
        - name: offset
          in: query
          description: オフセット（ページネーション用）
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
            example: 0
      responses:
        '200':
          description: 成功
          headers:
            X-Cache:
              $ref: '#/components/headers/CacheStatus'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntryListResponse'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /entries/hot:
    get:
      tags:
        - entries
      summary: 人気順エントリー一覧取得
      description: |
        指定日付のエントリーを人気順（bookmark_count DESC, posted_at DESC）で取得します。
        ブックマーク件数での閾値フィルタリングが可能です。
      operationId: getHotEntries
      parameters:
        - name: date
          in: query
          description: 取得対象日付（YYYYMMDD形式）
          required: true
          schema:
            type: string
            pattern: '^\d{8}$'
            example: "20250105"
        - name: min_users
          in: query
          description: 最低ブックマーク件数（5/10/50/100/500/1000など）
          required: false
          schema:
            type: integer
            minimum: 0
            default: 5
            example: 50
        - name: limit
          in: query
          description: 取得件数
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
            example: 25
        - name: offset
          in: query
          description: オフセット（ページネーション用）
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
            example: 0
      responses:
        '200':
          description: 成功
          headers:
            X-Cache:
              $ref: '#/components/headers/CacheStatus'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntryListResponse'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /rankings/yearly:
    get:
      tags:
        - rankings
      summary: 年次ランキング取得
      description: |
        指定年の「はてブ・オブ・ザ・イヤー」ランキングを取得します。
        ブックマーク件数上位順に並びます。
      operationId: getYearlyRanking
      parameters:
        - name: year
          in: query
          description: 対象年（YYYY形式）
          required: true
          schema:
            type: integer
            minimum: 2000
            maximum: 9999
            example: 2024
        - name: limit
          in: query
          description: 取得件数
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
            example: 100
      responses:
        '200':
          description: 成功
          headers:
            X-Cache:
              $ref: '#/components/headers/CacheStatus'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RankingResponse'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /rankings/monthly:
    get:
      tags:
        - rankings
      summary: 月次ランキング取得
      description: |
        指定年月の月間ランキングTOP 100を取得します。
        ブックマーク件数上位順に並びます。
      operationId: getMonthlyRanking
      parameters:
        - name: year
          in: query
          description: 対象年（YYYY形式）
          required: true
          schema:
            type: integer
            minimum: 2000
            maximum: 9999
            example: 2025
        - name: month
          in: query
          description: 対象月（MM形式、1-12）
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 12
            example: 12
        - name: limit
          in: query
          description: 取得件数
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 100
            example: 100
      responses:
        '200':
          description: 成功
          headers:
            X-Cache:
              $ref: '#/components/headers/CacheStatus'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RankingResponse'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /rankings/weekly:
    get:
      tags:
        - rankings
      summary: 週次ランキング取得
      description: |
        指定年・週の週間ランキングTOP 100を取得します。
        ブックマーク件数上位順に並びます。
      operationId: getWeeklyRanking
      parameters:
        - name: year
          in: query
          description: 対象年（YYYY形式）
          required: true
          schema:
            type: integer
            minimum: 2000
            maximum: 9999
            example: 2025
        - name: week
          in: query
          description: 対象週（1-53、ISO週番号）
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 53
            example: 1
        - name: limit
          in: query
          description: 取得件数
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 100
            example: 100
      responses:
        '200':
          description: 成功
          headers:
            X-Cache:
              $ref: '#/components/headers/CacheStatus'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RankingResponse'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API認証キー。データベースに登録されたキーと照合します。
        ヘッダーに `X-API-Key-ID: your-api-key-id` と `X-API-Key: your-api-key` の両方を含めてリクエストしてください。

        例:
        ```
        X-API-Key-ID: 9f8e7d6c-5b4a-3c2d-1e0f-123456789abc
        X-API-Key: hb_live_1234567890abcdef1234567890abcdef
        ```
  responses:
    UnauthorizedError:
      description: 認証エラー - APIキーまたはAPIキーIDが無効または未提供
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "UNAUTHORIZED"
            message: "Missing API key or key ID"
  headers:
    CacheStatus:
      description: キャッシュ判定（HIT/MISS）
      schema:
        type: string
        enum: [HIT, MISS]
  schemas:
    Entry:
      type: object
      description: はてなブックマークエントリー
      required:
        - id
        - title
        - url
        - posted_at
        - bookmark_count
        - tags
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: エントリーID
          example: "550e8400-e29b-41d4-a716-446655440000"
        title:
          type: string
          description: エントリーのタイトル
          example: "Go言語による高速なWeb API開発"
        url:
          type: string
          format: uri
          description: エントリーのURL
          example: "https://example.com/article/golang-web-api"
        posted_at:
          type: string
          format: date-time
          description: 記事の投稿日時（ISO 8601形式）
          example: "2025-01-05T10:30:00Z"
        bookmark_count:
          type: integer
          minimum: 0
          description: はてなブックマーク件数
          example: 150
        excerpt:
          type: string
          nullable: true
          description: 記事本文の抜粋
          example: "Go言語を使ったWeb API開発の実践的なテクニックを紹介します..."
        subject:
          type: string
          nullable: true
          description: RSSフィードのsubject（画面非表示、内部利用）
          example: "テクノロジー"
        tags:
          type: array
          description: エントリーに紐づくタグ一覧
          items:
            $ref: '#/components/schemas/EntryTag'
          example:
            - tag_id: "123e4567-e89b-12d3-a456-426614174000"
              tag_name: "Go"
              score: 95
            - tag_id: "223e4567-e89b-12d3-a456-426614174001"
              tag_name: "プログラミング"
              score: 85
        favicon_url:
          type: string
          format: uri
          description: Favicon URL（FaviconプロキシAPI）
          example: "/api/v1/favicons?domain=example.com"
        created_at:
          type: string
          format: date-time
          description: レコード作成日時（ISO 8601形式）
          example: "2025-01-05T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: レコード更新日時（ISO 8601形式）
          example: "2025-01-05T12:00:00Z"
    EntryTag:
      type: object
      description: エントリーに紐づくタグ（スコア付き）
      required:
        - tag_id
        - tag_name
        - score
      properties:
        tag_id:
          type: string
          format: uuid
          description: タグID
          example: "123e4567-e89b-12d3-a456-426614174000"
        tag_name:
          type: string
          maxLength: 100
          description: タグ名
          example: "Go"
        score:
          type: integer
          minimum: 0
          maximum: 100
          description: タグのスコア（Yahoo! キーフレーズ抽出APIから取得した重要度、0〜100）
          example: 95
    EntryListResponse:
      type: object
      description: エントリー一覧レスポンス
      required:
        - entries
        - total
        - limit
        - offset
      properties:
        entries:
          type: array
          description: エントリー一覧
          items:
            $ref: '#/components/schemas/Entry'
        total:
          type: integer
          minimum: 0
          description: 総件数（フィルタ適用後）
          example: 500
        limit:
          type: integer
          description: 1ページあたりの件数
          example: 25
        offset:
          type: integer
          description: オフセット
          example: 0
...
    RankingEntry:
      type: object
      description: ランキングエントリー
      required:
        - rank
        - entry
      properties:
        rank:
          type: integer
          minimum: 1
          description: 順位
          example: 1
        entry:
          $ref: '#/components/schemas/Entry'
    RankingResponse:
      type: object
      description: ランキングレスポンス
      required:
        - period_type
        - year
        - entries
        - total
      properties:
        period_type:
          type: string
          enum: [yearly, monthly, weekly]
          description: ランキング期間種別
          example: "monthly"
        year:
          type: integer
          description: 対象年
          example: 2025
        month:
          type: integer
          nullable: true
          description: 対象月（月次・週次ランキング時のみ）
          example: 12
        week:
          type: integer
          nullable: true
          description: 対象週（週次ランキング時のみ、ISO週番号）
          example: 1
        entries:
          type: array
          description: ランキングエントリー一覧
          items:
            $ref: '#/components/schemas/RankingEntry'
        total:
          type: integer
          minimum: 0
          description: 対象期間の総エントリー数
          example: 10000
...
    ErrorResponse:
      type: object
      description: エラーレスポンス
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: エラーコード
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: エラーメッセージ
          example: "Invalid parameter: date format must be YYYYMMDD"
        details:
          type: object
          nullable: true
          description: 詳細情報（オプション）
          additionalProperties: true
          example:
            field: "date"
            value: "20251301"
            constraint: "month must be 1-12"